Remove sudo
Currently we don't have sudo and most people have their MSYS2 roots
user-writable anyway, so let's completely remove the usage of sudo.  Most of the
patching is currently done in 0000-pacman-msysize.patch, these are just some new
occurrences.  Ideally this removal will be a completely separate patch in the
future.